Moved code for the signal attenuation mask to new file
