+++ echopype/mask/__init__.py
@@ -1,3 +1,5 @@
-from .api import apply_mask, frequency_differencing, get_attenuation_mask
+from .api import apply_mask, frequency_differencing
+from .mask_attenuated_signal import get_attenuation_mask
 
 __all__ = ["frequency_differencing", "apply_mask","get_attenuation_mask"]
+
