+++ echopype/mask/api.py
@@ -637,526 +637,3 @@ def frequency_differencing(
     da = da.assign_attrs({**mask_attrs, **{"history": history_attr}})
 
     return da
-
-
-def get_attenuation_mask(
-    source_Sv: Union[xr.Dataset, str, pathlib.Path],
-    mask_type: str = "ryan",
-    r0=180, r1=280, n=30, thr=6
-) -> xr.DataArray:
-    """
-    Create a mask based on the identified signal attenuations of Sv values at 38KHz. 
-    This method is based on:
-    Ryan et al. (2015) ‘Reducing bias due to noise and attenuation in 
-        open-ocean echo integration data’, ICES Journal of Marine Science,
-        72: 2482–2493.
-    and,
-    Ariza et al. (2022) 'Acoustic seascape partitioning through functional data analysis',
-    Journal of Biogeography, 00, 1– 15. https://doi.org/10.1111/jbi.14534
-    
-
-    Parameters
-    ----------
-    source_Sv: xr.Dataset or str or pathlib.Path
-        If a Dataset this value contains the Sv data to create a mask for,
-        else it specifies the path to a zarr or netcdf file containing
-        a Dataset. This input must correspond to a Dataset that has the
-        coordinate ``channel`` and variables ``frequency_nominal`` and ``Sv``.
-    mask_type: str with either "ryan" or "ariza" based on the prefered method for signal attenuation mask generation
-    Returns
-    -------
-    xr.DataArray
-        A DataArray containing the mask for the Sv data. Regions satisfying the thresholding
-        criteria are filled with ``True``, else the regions are filled with ``False``.
-
-    Raises
-    ------
-    ValueError
-        If neither ``ryan`` or ``azira`` are given
-
-    Notes
-    -----
-
-
-    Examples
-    --------
-
-    """
-    assert mask_type in ['ryan', 'ariza'], "mask_type must be either 'ryan' or 'ariza'"
-    
-    Sv = source_Sv['Sv'].values[0]
-    r = source_Sv['echo_range'].values[0,0]
-    if mask_type == "ryan":
-        mask = _get_attenuation_mask_ryan(Sv, r, r0=r0, r1=r1, n=n, thr=thr)
-    elif mask_type == "ariza":
-        mask = _get_attenuation_mask_ariza(Sv, r)
-    else:
-        raise ValueError(
-                "The provided mask_type must be ryan or ariza!"
-                )
-                
-    return_mask = xr.DataArray(mask,
-                                    dims=("ping_time", 
-                                          "range_sample"),
-                                    coords={"ping_time": source_Sv.ping_time,
-                                            "range_sample": source_Sv.range_sample}
-                                   )
-    return return_mask
-    
-    
-def _get_attenuation_mask_ryan(Sv, r, r0=400, r1=500, n=30, thr=8, start=0):
-    """
-    Locate attenuated signal and create a mask following the attenuated signal 
-    filter as in:
-        
-        Ryan et al. (2015) ‘Reducing bias due to noise and attenuation in 
-        open-ocean echo integration data’, ICES Journal of Marine Science,
-        72: 2482–2493.
-
-    Scattering Layers (SLs) are continuous high signal-to-noise regions with 
-    low inter-ping variability. But attenuated pings create gaps within SLs. 
-                                                 
-       attenuation                attenuation       ping evaluated
-    ______ V _______________________ V ____________.....V.....____________
-          | |   scattering layer    | |            .  block  .            |
-    ______| |_______________________| |____________...........____________|
-    
-    The filter takes advantage of differences with preceding and subsequent 
-    pings to detect and mask attenuation. A comparison is made ping by ping 
-    with respect to a block of the reference layer. The entire ping is masked 
-    if the ping median is less than the block median by a user-defined 
-    threshold value.
-    
-    Args:
-        Sv (float): 2D array with Sv data to be masked (dB). 
-        r (float):  1D array with range data (m).
-        r0 (int): upper limit of SL (m).
-        r1 (int): lower limit of SL (m).
-        n (int): number of preceding & subsequent pings defining the block.
-        thr (int): user-defined threshold value (dB).
-        start (int): ping index to start processing.
-        
-    Returns:
-        list: 2D boolean array with AS mask and 2D boolean array with mask
-              indicating where AS detection was unfeasible.
-    """
-    
-     # raise errors if wrong arguments
-    if r0>r1:
-        raise Exception('Minimum range has to be shorter than maximum range')
-    
-    # return empty mask if searching range is outside the echosounder range
-    if (r0>r[-1]) or (r1<r[0]):
-        mask  = np.zeros_like(Sv, dtype=bool) 
-        mask_ = np.zeros_like(Sv, dtype=bool) 
-        return mask, mask_ 
-    
-    # turn layer boundaries into arrays with length = Sv.shape[1]
-    r0 = np.ones(Sv.shape[1])*r0
-    r1 = np.ones(Sv.shape[1])*r1
-    
-    # start masking process    
-    mask_ = np.zeros(Sv.shape, dtype=bool)
-    mask = np.zeros(Sv.shape, dtype=bool) 
-    # find indexes for upper and lower SL limits
-    up = np.argmin(abs(r - r0))
-    lw = np.argmin(abs(r - r1))    
-    for j in range(start, len(Sv)):
-        
-        
-            # TODO: now indexes are the same at every loop, but future 
-            # versions will have layer boundaries with variable range
-            # (need to implement mask_layer.py beforehand!)
-        
-        # mask where AS evaluation is unfeasible (e.g. edge issues, all-NANs)
-        if (j-n<0) | (j+n>len(Sv)-1) | np.all(np.isnan(Sv[j,up:lw])):        
-            mask_[j, :] = True
-            
-        
-        # compare ping and block medians otherwise & mask ping if too different
-        else:
-            pingmedian  = _log(np.nanmedian(_lin(Sv[j, up:lw])))
-            blockmedian = _log(np.nanmedian(_lin(Sv[(j-n):(j+n),up:lw ])))
-            
-            if (pingmedian-blockmedian)<thr:            
-                mask[j, :] = True
-                
-    final_mask = np.logical_not(mask[start:, :] | mask_[start:, :]) 
-    return final_mask
-
-def _get_attenuation_mask_ariza(Sv, r, offset=20, thr=(-40,-35), m=20, n=50):
-    """
-    Mask attenuated pings by looking at seabed breaches.
-    
-    Ariza et al. (2022) 'Acoustic seascape partitioning through functional data analysis',
-    Journal of Biogeography, 00, 1– 15. https://doi.org/10.1111/jbi.14534
-    """
-    
-    # get ping array
-    p = np.arange(len(Sv))
-    # set to NaN shallow waters and data below the Sv threshold
-    Sv_ = Sv.copy()
-    Sv_[0:np.nanargmin(abs(r - offset)), :] = np.nan
-    Sv_[Sv_<-thr[0]] = np.nan
-    
-    # bin Sv
-    # TODO: update to 'twod' and 'full' funtions   
-    # DID    
-    irvals = np.linspace(p[0],p[-1],num=int((p[-1]-p[0])/n)+1)
-    jrvals = np.linspace(r[0],r[-1],num=int((r[-1]-r[0])/m)+1)
-    Sv_bnd, p_bnd, r_bnd = _twod(Sv_, p, r, irvals, jrvals, operation='mean')[0:3]
-    print(r_bnd)
-    print(p_bnd)
-    print("originals:")
-    print(r)
-    print(p)
-    Sv_bnd = _full(Sv_bnd, r_bnd, p_bnd, r, p)
-    
-    # label binned Sv data features
-    Sv_lbl = label(~np.isnan(Sv_bnd))
-    labels = np.unique(Sv_lbl)
-    labels = np.delete(labels, np.where(labels==0))
-    
-    # list the median values for each Sv feature
-    val = []
-    for lbl in labels:
-        val.append(_log(np.nanmedian(_lin(Sv_bnd[Sv_lbl==lbl]))))
-    
-    # keep the feature with a median above the Sv threshold (~seabed)
-    # and set the rest of the array to NaN
-    if val:
-        if np.nanmax(val)>thr[1]:
-            labels = labels[val!=np.nanmax(val)]
-            for lbl in labels:
-                Sv_bnd[Sv_lbl==lbl] = np.nan
-        else:
-            Sv_bnd[:] = np.nan
-    else:
-        Sv_bnd[:] = np.nan
-        
-    # remove everything in the original Sv array that is not seabed
-    Sv_sb = Sv.copy()
-    Sv_sb[np.isnan(Sv_bnd)] = np.nan
-    
-    # compute the percentile 90th for each ping, at the range at which 
-    # the seabed is supposed to be.    
-    seabed_percentile = _log(np.nanpercentile(_lin(Sv_sb), 95, axis=0))
-    
-    # get mask where this value falls bellow a Sv threshold (seabed breaches)
-    mask = seabed_percentile<thr[0]
-    mask = np.tile(mask, [len(Sv), 1])    
-    
-    return mask
-    
-    
-def _lin(variable):
-    """
-    Turn variable into the linear domain.     
-    
-    Args:
-        variable (float): array of elements to be transformed.    
-    
-    Returns:
-        float:array of elements transformed
-    """
-    
-    lin    = 10**(variable/10)
-    return   lin  
-
-def _log(variable):
-    """
-    Turn variable into the logarithmic domain. This function will return -999
-    in the case of values less or equal to zero (undefined logarithm). -999 is
-    the convention for empty water or vacant sample in fisheries acoustics. 
-    
-    Args:
-        variable (float): array of elements to be transformed.    
-    
-    Returns:
-        float: array of elements transformed
-    """
-    
-    # convert variable to float array
-    back2single = False
-    back2list   = False
-    back2int    = False    
-    if not isinstance(variable, np.ndarray):
-        if isinstance(variable, list):
-            variable  = np.array(variable)
-            back2list = True
-        else:
-            variable    = np.array([variable])
-            back2single = True            
-    if variable.dtype=='int64':
-        variable   = variable*1.0
-        back2int = True
-    
-    # compute logarithmic value except for zeros values, which will be -999 dB    
-    mask           = np.ma.masked_less_equal(variable, 0).mask
-    variable[mask] = np.nan
-    log            = 10*np.log10(variable)
-    log[mask]      = -999
-    
-    # convert back to original data format and return
-    if back2int:
-        log = np.int64(log)
-    if back2list:
-        log = log.tolist()
-    if back2single:
-        log = log[0]       
-    return log
-    
-    
-    
-def _full(datar, irvals, jrvals, idim, jdim):
-    """
-    Turn resampled data back to full resolution, according to original i and j
-    full resolution dimensions.
-    
-    Args:
-        datar  (float): 2D array with resampled data.
-        irvals (float): 1D array with i resampling intervals.
-        jdimr  (float): 1D array with i resampling intervals.
-        idim   (float): 1D array with full resolution i axis.
-        jdim   (float): 1D array with full resolution j axis.
-        
-    Returns:
-        float: 2D array with data resampled at full resolution.
-        bool : 2D array with mask indicating valid values in data resampled
-               at full resolution.
-    """
-    
-    
-    # check for appropiate inputs
-    if len(irvals)<2:
-        raise Exception('i resampling interval length must be >2')
-    if len(jrvals)<2:
-        raise Exception('j resampling interval length must be >2')
-    if len(datar)+1<len(irvals):
-        raise Exception('i resampling intervals length can\'t exceed data height + 1')
-    if len(datar[0])+1<len(jrvals):
-        raise Exception('j resampling intervals length can\'t exceed data width + 1') 
-        
-    # get i/j axes from i/j dimensions and i/j intervals
-    iax   = np.arange(len(idim))
-    jax   = np.arange(len(jdim))
-    iaxrs = dim2ax(idim, iax, irvals)    
-    jaxrs = dim2ax(jdim, jax, jrvals)
-    
-    # check whether i/j resampled axes and i/j intervals are different
-    idiff = True
-    if len(iaxrs)==len(iax):
-        if (iaxrs==iax).all():
-            idiff = False
-    jdiff = True
-    if len(jaxrs)==len(jax):
-        if (jaxrs==jax).all():
-            jdiff = False
-    
-    # preallocate full resolution data array 
-    data = np.zeros((len(iax), len(jax)))*np.nan
-        
-    # if i/j axes are different, resample back to full along i/j dimensions
-    if idiff&jdiff:
-        for i in range(len(iaxrs)-1):            
-            idx = np.where((iaxrs[i]<=iax) & (iax<iaxrs[i+1]))[0]
-            for j in range(len(jaxrs)-1):        
-                jdx = np.where((jaxrs[j]<=jax) & (jax<jaxrs[j+1]))[0]
-                if idx.size*jdx.size > 0:
-                    data[idx[0]:idx[-1]+1, jdx[0]:jdx[-1]+1]= datar[i,j]
-    
-    # if only i axis is different, resample back to full along i dimension
-    elif idiff & (not jdiff):
-        for i in range(len(iaxrs)-1):            
-            idx = np.where((iaxrs[i]<=iax) & (iax<iaxrs[i+1]))[0]
-            if idx.size>0:
-                data[idx[0]:idx[-1]+1, :]= datar[i, :]
-        
-    # if only j axis is different, resample back to full along j dimension
-    elif (not idiff) & jdiff:        
-        for j in range(len(jaxrs)-1):        
-            jdx = np.where((jaxrs[j]<=jax) & (jax<jaxrs[j+1]))[0]
-            if jdx.size > 0:
-                data[:, jdx[0]:jdx[-1]+1]= datar[:, j].reshape(-1,1)
-        
-    # if i/j resampled & i/j full are the same, data resampled & data are equal
-    else:
-        warnings.warn("Array already at full resolution!", RuntimeWarning)
-        data= datar.copy()
-    
-    # get mask indicating where data couldn't be resampled back
-    mask_ = np.zeros_like(data, dtype=bool)
-    i1= np.where(iax<iaxrs[-1])[0][-1] + 1
-    j1= np.where(jax<jaxrs[-1])[0][-1] + 1
-    if idiff:
-        mask_[i1:,   :] = True
-    if jdiff:
-        mask_[  :, j1:] = True
-    
-    return data, mask_
-    
-    
-def _twod(data, idim, jdim, irvals, jrvals, log=False, operation='mean'):
-    """
-    Resample down an array along the two dimensions, i and j.
-    
-    Args:
-        data   (float): 2D array with data to be resampled.
-        idim   (float): i vertical dimension.
-        jdim   (float): j horizontal dimension.
-        irvals (float): i resampling intervals for i vertical dimension.
-        jrvals (float): j resampling intervals for j horizontal dimension.
-        log    (bool ): if True, data is considered logarithmic and it will 
-                        be converted to linear during calculations.
-        operation(str): type of resampling operation. Accepts "mean" or "sum".
-                    
-    Returns:
-        float: 2D resampled data array
-        float: 1D resampled i vertical dimension
-        float: 1D resampled j horizontal dimension
-        float: 2D array with percentage of valid samples included on each
-               resampled cell.
-    """ 
-    
-    # check for appropiate inputs
-    if len(irvals)<2:
-        raise Exception('length of i resampling intervals must be  >2')
-    if len(jrvals)<2:
-        raise Exception('length of j resampling intervals must be >2')
-    if len(data)!=len(idim):
-        raise Exception('data height and idim length must be the same')
-    if len(data[0])!=len(jdim):
-        raise Exception('data width and jdim length must be the same')        
-    for irval, jrval in zip(irvals, jrvals):
-        if (irval<idim[0]) | (idim[-1]<irval):
-            raise Exception('i resampling intervals must be within idim range')
-        if (jrval<jdim[0]) | (jdim[-1]<jrval):
-            raise Exception('j resampling intervals must be within jdim range')
-        
-    # convert data to linear, if logarithmic
-    if log is True:
-        data = _lin(data)
-    
-    # get i/j axes from i/j dimensions and i/j intervals
-    iax   = np.arange(len(idim))
-    jax   = np.arange(len(jdim))
-    iaxrs = dim2ax(idim, iax, irvals)    
-    jaxrs = dim2ax(jdim, jax, jrvals)
-    
-    # declare new array to allocate resampled values, and new array to
-    # alllocate the percentage of values used for resampling
-    datar     = np.zeros((len(iaxrs)-1, len(jaxrs)-1))*np.nan
-    percentage = np.zeros((len(iaxrs)-1, len(jaxrs)-1))*np.nan
-    
-    # iterate along-range
-    for i in range(len(iaxrs)-1):
-        
-        # get i indexes to locate the samples for the binning operation
-        idx0     = np.where(iax-iaxrs[i+0]<=0)[0][-1]        
-        idx1     = np.where(iaxrs[i+1]-iax> 0)[0][-1]
-        idx      = np.arange(idx0, idx1+1)
-        
-        # get i weights as the sum of the sample proportions taken
-        iweight0 = 1 - abs(iax[idx0]-iaxrs[i+0])
-        iweight1 = abs(iax[idx1]-iaxrs[i+1])
-        if len(idx)>1:
-            iweights = np.r_[iweight0, np.ones(len(idx)-2), iweight1]
-        else:
-            iweights = np.array([iweight0-1 + iweight1])
-        
-        # iterate along-pings
-        for j in range(len(jaxrs)-1):
-            
-            # get j indexes to locate the samples for the binning operation
-            jdx0     = np.where(jax-jaxrs[j+0]<=0)[0][-1]        
-            jdx1     = np.where(jaxrs[j+1]-jax> 0)[0][-1]
-            jdx      = np.arange(jdx0, jdx1+1)
-            
-            # get j weights as the sum of the sample proportions taken
-            jweight0 = 1 - abs(jax[jdx0]-jaxrs[j+0])
-            jweight1 = abs(jax[jdx1]-jaxrs[j+1])
-            if len(jdx)>1:
-                jweights = np.r_[jweight0, np.ones(len(jdx)-2), jweight1]
-            else:
-                jweights = np.array([jweight0-1 + jweight1])
-                            
-            # get data and weight 2D matrices for the binning operation
-            d = data[idx[0]:idx[-1]+1, jdx[0]:jdx[-1]+1]
-            w = np.multiply.outer(iweights, jweights)
-                      
-            # if d is an all-NAN array, return NAN as the weighted operation
-            # and zero as the percentage of valid numbers used for binning
-            if np.isnan(d).all():
-                datar    [i, j] = np.nan
-                percentage[i, j] = 0
-            
-            #compute weighted operation & percentage of valid numbers otherwise
-            else:                    
-                w_ = w.copy()
-                w_[np.isnan(d)] = np.nan
-                if operation=='mean':
-                    datar   [i, j]  = np.nansum(d*w_)/np.nansum(w_)
-                elif operation=='sum':
-                    datar   [i, j]  = np.nansum(d*w_)
-                else:
-                    raise Exception('Operation not recognised')                        
-                percentage[i, j]  = np.nansum(  w_)/np.nansum(w )*100                        
-    
-    # convert back to logarithmic, if data was logarithmic
-    if log is True:
-        datar = _log(datar)
-    
-    # get resampled dimensions from resampling intervals
-    idimr = irvals[:-1]
-    jdimr = jrvals[:-1]
-    
-    # return data
-    return datar, idimr, jdimr, percentage
-    
-def dim2ax(dim, ax, dimrs):
-    """
-    It gives you a new resampled axis based on a known dimension/axis pair, and 
-    a new resampled dimesion.
-    
-    Args:
-        dim   (float): 1D array with original dimension.
-        ax    (int  ): 1D array with original axis.
-        dimrs (float): 1D array with resampled dimension.
-    
-    Returns:
-        float: 1D array with resampled axis.
-    
-    Notes:
-        Dimension refers to range, time, latitude, distance, etc., and axis
-        refer to dimension indexes such as sample or ping number.
-        
-    Example:
-                                                         (resampled)
-        seconds dimension | ping axis       seconds dimension | ping axis
-        ------------------·----------       ------------------·----------
-                       0  | 0                             0   | 0
-                       2  | 1                             3   | 1.5
-                       4  | 2          ==>                6   | 3.0
-                       6  | 3                             9   | 4.5
-                       8  | 4                             -   | -
-                      10  | 4                             -   | -
-    """
-    
-    # check that resampled dimension doesn't exceed the limits
-    # of the original one
-    if (dimrs[0]<dim[0]) | (dimrs[-1]>dim[-1]):
-        raise Exception('resampling dimension can not exceed ' +
-                        'the original dimension limits') 
-        
-    # convert variables to float64 if they are in datetime64 format
-    epoch = np.datetime64('1970-01-01T00:00:00')    
-    if 'datetime64' in str(dim[0].dtype):        
-        dim = np.float64(dim-epoch)        
-    if 'datetime64' in str(dimrs[0].dtype):
-        dimrs = np.float64(dimrs-epoch)
-    
-    # get interpolated new y variable    
-    f    = interp1d(dim, ax)
-    axrs = f(dimrs)
-            
-    return axrs
